feat: add authentication and authorization

+++ api/routers/charging_station_router.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from typing import List
 from pydantic import UUID4
@@ -7,6 +7,7 @@ from ..crud import crud
 from ..schemas import schemas
 from ..database.database import get_db
 from ..models.models import CurrentTypeEnum
+from ..utils.auth import get_current_user
 
 logger = logging.getLogger(__name__)
 
@@ -22,9 +23,17 @@ router = APIRouter()
 )
 def create_charging_station(
         charging_station_data: schemas.ChargingStationCreate,
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
 ):
     logger.info(f"Attempting to create a new charging station with data: {charging_station_data}")
+    if not current_user:
+        logger.error("Unauthorized attempt to create a charging station.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to create a charging station.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.create_charging_station(db=db, charging_station_data=charging_station_data)
     logger.info(f"Successfully created a charging station with ID: {result.id}")
     return result
@@ -37,8 +46,19 @@ def create_charging_station(
     tags=["Charging Stations"],
     summary="Read a specific charging station providing its UUID"
 )
-def read_charging_station(charging_station_id: UUID4, db: Session = Depends(get_db)):
+def read_charging_station(
+        charging_station_id: UUID4,
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
+):
     logger.info(f"Fetching a charging station with ID: {charging_station_id}")
+    if not current_user:
+        logger.error("Unauthorized attempt to retrieve a charging station.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to retrieve a charging station.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.get_charging_station(db=db, charging_station_id=charging_station_id)
     logger.info(f"Successfully retrieved a charging station with ID: {charging_station_id}")
     return result
@@ -59,13 +79,21 @@ def read_charging_station_list(
         firmware_version: str | None = None,
         skip: int = 0,
         limit: int = 100,
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
 ):
     logger.info(
         f"Fetching a list of charging stations with filters: plug_count={plug_count}, min_efficiency={min_efficiency}, "
         f"max_efficiency={max_efficiency}, current_type={current_type}, firmware_version={firmware_version}, "
         f"skip={skip}, limit={limit}."
     )
+    if not current_user:
+        logger.error("Unauthorized attempt to retrieve charging stations.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to retrieve charging stations.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.get_charging_station_list(
         plug_count=plug_count,
         min_efficiency=min_efficiency,
@@ -90,12 +118,20 @@ def read_charging_station_list(
 def update_charging_station(
         charging_station_id: UUID4,
         charging_station_data: schemas.ChargingStationCreate,
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
 ):
     logger.info(
         f"Attempting to update a charging station with ID: {charging_station_id} with data: "
         f"{charging_station_data}."
     )
+    if not current_user:
+        logger.error("Unauthorized attempt to update a charging station.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to update a charging station.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.update_charging_station(
         db=db,
         charging_station_id=charging_station_id,
@@ -111,8 +147,19 @@ def update_charging_station(
     tags=["Charging Stations"],
     summary="Delete a specific charging station providing its UUID"
 )
-def delete_charging_station(charging_station_id: UUID4, db: Session = Depends(get_db)):
+def delete_charging_station(
+        charging_station_id: UUID4,
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
+):
     logger.info(f"Trying to delete a charging station with ID: {charging_station_id}.")
+    if not current_user:
+        logger.error("Unauthorized attempt to delete a charging station.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to delete a charging station.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.delete_charging_station(db=db, charging_station_id=charging_station_id)
     logger.info(f"Successfully deleted a charging station with ID: {charging_station_id}.")
     return result
