+++ api/routers/charging_station_type_router.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from typing import List
 from pydantic import UUID4
@@ -7,6 +7,7 @@ from ..crud import crud
 from ..schemas import schemas
 from ..database.database import get_db
 from ..models.models import CurrentTypeEnum
+from ..utils.auth import get_current_user
 
 logger = logging.getLogger(__name__)
 
@@ -22,9 +23,17 @@ router = APIRouter()
 )
 def create_charging_station_type(
         charging_station_type_data: schemas.ChargingStationTypeCreate,
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
 ):
     logger.info(f"Attempting to create a charging station type with data: {charging_station_type_data}.")
+    if not current_user:
+        logger.error("Unauthorized attempt to create a charging station type.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to create a charging station type.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.create_charging_station_type(db=db, charging_station_type_data=charging_station_type_data)
     logger.info(f"Successfully created a charging station type with ID: {result.id}.")
     return result
@@ -37,8 +46,19 @@ def create_charging_station_type(
     tags=["Charging Station Types"],
     summary="Read a specific charging station type providing its UUID"
 )
-def read_charging_station_type(charging_station_type_id: UUID4, db: Session = Depends(get_db)):
+def read_charging_station_type(
+        charging_station_type_id: UUID4,
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
+):
     logger.info(f"Fetching a charging station type with ID: {charging_station_type_id}.")
+    if not current_user:
+        logger.error("Unauthorized attempt to retrieve a charging station type.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to retrieve a charging station type.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.get_charging_station_type(db=db, charging_station_type_id=charging_station_type_id)
     logger.info(f"Successfully retrieved a charging station type with ID: {charging_station_type_id}.")
     return result
@@ -58,13 +78,21 @@ def read_charging_station_type_list(
         current_type: CurrentTypeEnum | None = None,
         skip: int = 0,
         limit: int = 100,
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
 ):
     logger.info(
         f"Fetching a list of charging station types with filters: plug_count={plug_count}, "
         f"min_efficiency={min_efficiency}, max_efficiency={max_efficiency}, current_type={current_type}, "
         f"skip={skip}, limit={limit}."
     )
+    if not current_user:
+        logger.error("Unauthorized attempt to retrieve charging station types.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to retrieve charging station types.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.get_charging_station_type_list(
         plug_count=plug_count,
         min_efficiency=min_efficiency,
@@ -88,12 +116,20 @@ def read_charging_station_type_list(
 def update_charging_station_type(
         charging_station_type_id: UUID4,
         charging_station_type_data: schemas.ChargingStationTypeCreate,
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
 ):
     logger.info(
         f"Attempting to update a charging station type with ID: {charging_station_type_id} "
         f"with data: {charging_station_type_data}."
     )
+    if not current_user:
+        logger.error("Unauthorized attempt to update a charging station type.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to update a charging station type.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.update_charging_station_type(
         db=db,
         charging_station_type_id=charging_station_type_id,
@@ -109,8 +145,19 @@ def update_charging_station_type(
     tags=["Charging Station Types"],
     summary="Delete a specific charging station type providing its UUID"
 )
-def delete_charging_station_type(charging_station_type_id: UUID4, db: Session = Depends(get_db)):
+def delete_charging_station_type(
+        charging_station_type_id: UUID4,
+        db: Session = Depends(get_db),
+        current_user: schemas.User = Depends(get_current_user)
+):
     logger.info(f"Attempting to delete a charging station type with ID: {charging_station_type_id}.")
+    if not current_user:
+        logger.error("Unauthorized attempt to delete a charging station type.")
+        raise HTTPException(
+            status_code=401,
+            detail="Not authorized to delete a charging station type.",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
     result = crud.delete_charging_station_type(db=db, charging_station_type_id=charging_station_type_id)
     logger.info(f"Successfully deleted a charging station type with ID: {charging_station_type_id}.")
     return result
